requirements.txt
torchmetrics
numpy==1.26.4
pandas
tqdm
matplotlib
librosa>=0.10.0

soundfile

torchaudio==2.4.0

omegaconf>=2.3.0

wandb
comet_ml

hydra-core

black
isort
pre-commit